D1 mini: Measure alternating current using INA122 and ASM010
The selfconstruction shield D1_INA122 can be used to amplify and rectify small alternating voltages. For this purpose a precision instrument amplifier INA122P and a rectifier diode BAT48 are used.
If an ASM010 current transformer with shunt resistor is used for the AC current measurement, the value can be measured using the Do-It-Yourself Shield D1_INA122 via the Analog In input A0 of the D1 mini.